requirements.txt
streamlit
lorem-text
crewai
crewai_tools
newspaper3k
lxml_html_clean
